+++ cmd/psweb/ln/common.go
@@ -36,6 +36,15 @@ type ForwardingStats struct {
 	AssistedFeeSat6m  uint64
 }
 
+type ChanneInfo struct {
+	ChannelId     string
+	LocalBalance  uint64
+	RemoteBalance uint64
+	FeeRate       uint64
+	FeeBase       uint64
+	Active        bool
+}
+
 func toSats(amount float64) int64 {
 	return int64(float64(100000000) * amount)
 }